修复migrations

+++ database/migrations/2017_10_17_200205_create_ims_yz_deduction_table.php
@@ -13,7 +13,7 @@ class CreateImsYzDeductionTable extends Migration
      */
     public function up()
     {
-        if (!\Schema::hasColumn('yz_deduction', 'code')) {
+        if (Schema::hasTable('yz_deduction') && !\Schema::hasColumn('yz_deduction', 'code')) {
             Schema::drop('yz_deduction');
         }
         if (!Schema::hasTable('yz_deduction')) {
